+++ src/main.tsx
@@ -9,6 +9,7 @@ import { HeatmapProvider } from "./context/heatmap/heatmap.context";
 import "./localization/i18n";
 import { useContext } from "react";
 import { IHeatmapView } from "src/types";
+import { mergeTrackerData } from "./utils/core";
 
 declare global {
   interface Window {
@@ -42,18 +43,25 @@ const DEFAULT_SETTINGS: TrackerSettings = {
     [IHeatmapView.Donation]: true,
     [IHeatmapView.HeatmapTracker]: true,
     [IHeatmapView.HeatmapTrackerStatistics]: true,
+    [IHeatmapView.Legend]: true,
   },
 };
 
-const DEFAULT_TRACKER_DATA: TrackerData = {
+export const DEFAULT_TRACKER_DATA: TrackerData = {
   year: new Date().getFullYear(),
   entries: [
     { date: "1900-01-01", color: "#7bc96f", intensity: 5, content: "" },
   ],
   showCurrentDayBorder: true,
+  intensityConfig: {
+    scaleStart: undefined,
+    scaleEnd: undefined,
+    defaultIntensity: 4,
+    showOutOfRange: true,
+  },
+  intensityScaleStart: undefined,
+  intensityScaleEnd: undefined,
   defaultEntryIntensity: 4,
-  intensityScaleStart: 1,
-  intensityScaleEnd: 5,
   colorScheme: {
     paletteName: "default",
   },
@@ -68,7 +76,7 @@ export default class HeatmapTracker extends Plugin {
 
     window.renderHeatmapTracker = (
       el: HTMLElement,
-      trackerData: TrackerData
+      trackerData: TrackerData = DEFAULT_TRACKER_DATA
     ): void => {
       const container = el.createDiv({
         cls: "heatmap-tracker-container",
@@ -81,7 +89,7 @@ export default class HeatmapTracker extends Plugin {
         <StrictMode>
           <AppContext.Provider value={this.app}>
             <HeatmapProvider
-              trackerData={{ ...DEFAULT_TRACKER_DATA, ...trackerData }}
+              trackerData={mergeTrackerData(DEFAULT_TRACKER_DATA, trackerData)}
               settings={this.settings}
             >
               <ReactApp />