+++ src/utils/tabs.tsx
@@ -1,15 +1,15 @@
-import { HeatmapIcon } from "../components/icons/HeatmapIcon";
-import { StatisticsIcon } from "../components/icons/StatisticsIcon";
-// import { MenuIcon } from "../components/icons/MenuIcon";
-import { DocumentationIcon } from "../components/icons/DocumentationIcon";
+import { HeatmapIcon } from "src/components/icons/HeatmapIcon";
+import { StatisticsIcon } from "src/components/icons/StatisticsIcon";
+import { DocumentationIcon } from "src/components/icons/DocumentationIcon";
 import { ReactNode } from "react";
-import { HandCoinsIcon } from "../components/icons/HandCoinsIcon";
+import { HandCoinsIcon } from "src/components/icons/HandCoinsIcon";
 import { IHeatmapView } from "src/types";
+import { LegendIcon } from "src/components/icons/LegendIcon";
 
 export const TabIconForView: Record<IHeatmapView, ReactNode> = {
   [IHeatmapView.HeatmapTracker]: <HeatmapIcon />,
   [IHeatmapView.HeatmapTrackerStatistics]: <StatisticsIcon />,
-  // [IHeatmapView.HeatmapMenu]: <MenuIcon />,
   [IHeatmapView.Documentation]: <DocumentationIcon />,
   [IHeatmapView.Donation]: <HandCoinsIcon />,
+  [IHeatmapView.Legend]: <LegendIcon />,
 };