+++ CHANGELOG.md
@@ -6,11 +6,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 
 ---
 
+## [1.12.0] - 2025-01-06
+
+### Added
+- Add legend view.
+- Add `showOutOfRange` property.
+- Add tests for utils.
+
+## Changed
+- Update intensity calculation.
+- Change font-size of documentation view.
+
 ## [1.11.1] - 2025-01-01
 
 ### Fixed
 - Fix default year (2024 -> 2025).
-
 
 ## [1.11.0] - 2024-12-28